+++ MatCalcitePlugin/src/com/github/vlsi/mat/calcite/functions/CollectionsFunctions.java
@@ -15,19 +15,19 @@ import org.apache.calcite.schema.ScalarFunction;
 import org.apache.calcite.schema.impl.ReflectiveFunctionBase;
 import org.eclipse.mat.SnapshotException;
 import org.eclipse.mat.inspections.collectionextract.CollectionExtractionUtils;
+import org.eclipse.mat.inspections.collectionextract.ExtractedCollection;
 import org.eclipse.mat.inspections.collectionextract.ExtractedMap;
 import org.eclipse.mat.snapshot.model.IObject;
 
 import java.lang.reflect.Method;
-import java.util.Collections;
-import java.util.HashMap;
-import java.util.Map;
+import java.util.*;
 
-public class MapFunctions extends HeapFunctionsBase {
-    public static class MapFunction extends ReflectiveFunctionBase implements ScalarFunction, ImplementableFunction {
-        private final CallImplementor implementor;
+public class CollectionsFunctions extends HeapFunctionsBase {
 
-        public MapFunction(Method method) {
+    public abstract static class BaseImplementableFunction extends ReflectiveFunctionBase implements ScalarFunction, ImplementableFunction {
+        final CallImplementor implementor;
+
+        BaseImplementableFunction(Method method) {
             super(method);
             implementor = RexImpTable.createImplementor(new ReflectiveCallNotNullImplementor(method), NullPolicy.NONE, false);
         }
@@ -36,6 +36,12 @@ public class MapFunctions extends HeapFunctionsBase {
         public CallImplementor getImplementor() {
             return implementor;
         }
+    }
+
+    public static class MapFunction extends BaseImplementableFunction {
+        MapFunction(Method method) {
+            super(method);
+        }
 
         @Override
         public RelDataType getReturnType(RelDataTypeFactory relDataTypeFactory) {
@@ -43,9 +49,21 @@ public class MapFunctions extends HeapFunctionsBase {
         }
     }
 
-    public static Multimap<String, MapFunction> createAll() {
-        ImmutableMultimap.Builder<String, MapFunction> builder = ImmutableMultimap.builder();
-        builder.put("asMap", new MapFunction(findMethod(MapFunctions.class, "asMap")));
+    public static class MultiSetFunction extends BaseImplementableFunction {
+        MultiSetFunction(Method method) {
+            super(method);
+        }
+
+        @Override
+        public RelDataType getReturnType(RelDataTypeFactory relDataTypeFactory) {
+            return relDataTypeFactory.createMultisetType(relDataTypeFactory.createJavaType(HeapReference.class), -1);
+        }
+    }
+
+    public static Multimap<String, ScalarFunction> createAll() {
+        ImmutableMultimap.Builder<String, ScalarFunction> builder = ImmutableMultimap.builder();
+        builder.put("asMap", new MapFunction(findMethod(CollectionsFunctions.class, "asMap")));
+        builder.put("asMultiSet", new MultiSetFunction(findMethod(CollectionsFunctions.class, "asMultiSet")));
         return builder.build();
     }
 
@@ -73,5 +91,28 @@ public class MapFunctions extends HeapFunctionsBase {
         } catch (SnapshotException e) {
             throw new RuntimeException("Unable to extract map from " + r, e);
         }
+    }
+
+    public static List asMultiSet(Object r) {
+        HeapReference ref = ensureHeapReference(r);
+        if (ref == null) {
+            return null;
+        }
+
+        try {
+            ExtractedCollection extractedCollection = CollectionExtractionUtils.extractList(ref.getIObject());
+            if (extractedCollection == null) {
+                return Collections.emptyList();
+            } else {
+                List<HeapReference> result = new ArrayList<>();
+                for(IObject entry :extractedCollection) {
+                    result.add((HeapReference)resolveReference(entry));
+                }
+                return result;
+            }
+        } catch (SnapshotException e) {
+            throw new RuntimeException("Unable to extract collection from " + r, e);
+        }
+
     }
 }